Refactor Twig extension
