+++ templates/post/by_date.html.twig
@@ -38,8 +38,8 @@
 
 {%- endblock -%}
 {%- block javascripts -%}
+    {{- parent() -}}
     <script>
         require(['bbc-datepicker']);
     </script>
-    {{- parent() -}}
 {%- endblock -%}
